+++ lib/arel/extensions.rb
@@ -19,6 +19,7 @@ require 'arel/extensions/cross_join'
 require 'arel/extensions/lateral'
 require 'arel/extensions/range_function'
 require 'arel/extensions/with_ordinality'
+require 'arel/extensions/table'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
