+++ lib/arel/extensions/nodes.rb
@@ -4,31 +4,6 @@
 
 module Arel
   module Nodes
-    Arel::Table.class_eval do
-      # postgres only: https://www.postgresql.org/docs/9.5/sql-select.html
-      attr_accessor :only
-      # postgres only: https://www.postgresql.org/docs/9.5/ddl-schemas.html
-      attr_accessor :schema_name
-      # postgres only: https://www.postgresql.org/docs/9.1/catalog-pg-class.html
-      attr_accessor :relpersistence
-
-      alias_method :old_initialize, :initialize
-      def initialize(
-        name,
-        as: nil,
-        type_caster: nil,
-        only: false,
-        schema_name: nil,
-        relpersistence: 'p'
-      )
-        @only = only
-        @schema_name = schema_name
-        @relpersistence = relpersistence
-
-        old_initialize(name, as: as, type_caster: type_caster)
-      end
-    end
-
     # postgres only: https://www.postgresql.org/docs/9.2/sql-expressions.html
     class Row < Arel::Nodes::Unary
       attr_reader :row_format
